Move all go code into the `proxy-init` directory
In preparation for introducing components other than the `proxy-init`
binary, this change moves all of the current Go code into a
`./proxy-init` directory.

This changes the relevant Just recipes to be prefixed with
`proxy-init-`.
