+++ proxy-init/cmd/root.go
@@ -7,8 +7,8 @@ import (
 
 	log "github.com/sirupsen/logrus"
 
-	"github.com/linkerd/linkerd2-proxy-init/iptables"
-	"github.com/linkerd/linkerd2-proxy-init/ports"
+	"github.com/linkerd/linkerd2-proxy-init/proxy-init/pkg/iptables"
+	"github.com/linkerd/linkerd2-proxy-init/proxy-init/pkg/ports"
 	"github.com/spf13/cobra"
 )
